+++ modules/axfs/src/lib.rs
@@ -3,3 +3,6 @@
 pub mod memfs;
 
 pub use axvfs::{FileType, InodeId, Metadata};
+
+#[cfg(test)]
+extern crate std;